+++ principles/src/main/java/com/thecollinsbyte/principles/SOLID/openclosed/Operation.java
@@ -1,5 +1,5 @@
 package com.thecollinsbyte.principles.SOLID.openclosed;
 
 public interface Operation {
-    int calculate(int num1, int num2);
+    double calculate(double num1, double num2);
 }
