+++ principles/src/test/java/com/thecollinsbyte/principles/SOLID/openclosed/CalculatorTest.java
@@ -9,7 +9,7 @@ class CalculatorTest {
     @Test
     void addOperation() {
         AddOperation addOperation = new AddOperation();
-        int result = addOperation.calculate(10, 10);
+        double result = addOperation.calculate(10, 10);
 
         assertEquals(20, result);
     }
@@ -17,7 +17,7 @@ class CalculatorTest {
     @Test
     void subtractOperation() {
         SubtractOperation subtractOperation = new SubtractOperation();
-        int result = subtractOperation.calculate(10, 10);
+        double result = subtractOperation.calculate(10, 10);
 
         assertEquals(0, result);
     }
